chore: shared fetch.py with yml config for arcgis (#26)
ArcGIS feeds have a repeatable `fetch` pattern, and @obra observed that
we could make it easier to add new feeds by re-using a single runner.

Augment `vaccine-feed-ingest` to accept `.yml` configuration files if an
executable is not provided (for the `fetch` stage), and use that config
to run a shared ArcGIS scraper. This setup is designed so that the
`vaccine-feed-ingest` cli commands continue to operate as expected even
when a config file is found rather than an executable runner. Additionally,
while ArcGIS is the only shared scraper at the moment, this setup should
be able to support more shared scrapers with minimal refactoring.