Normalize Arizona arcgis vaccine data (fixes #59) (#117)
* Normalize Arizona arcgis data (fixes #59)

Partially adapted from the Alaska normalizer.

The main complexity is parsing site hours, which are in a very wide variety of time formats; it seems like a value is produced for almost all of them. I wasn't sure how deeply I should go into handling special cases versus keeping the normalizer simple; currently it's able to parse almost all of the times, but is fairly complicated and might not be robust. (Some sites have the hours listed for particular days as the string "BY APPOINTMENT ONLY" or "7:00 AM-5:00PM for second dose vaccine ONLY"; these are theoretically parseable, but it's not clear how they should be parsed here. The time parsing just falls back to `None` if it can't figure out what's going on.)

* Fix linting/open and close date/phone number parsing

* fix: update az fetched from url

Co-authored-by: eliblock <3347571+eliblock@users.noreply.github.com>