chore(deps)!: bump vaccine_feed_ingest_schema to v1.0.0 (#295)
Update vaccine_feed_ingest_schema to v1.0.0 ([release notes](https://github.com/CAVaccineInventory/vaccine-feed-ingest-schema/releases/tag/1.0.0)).

This adds significant validation to the schema, which will make it must
easier to extract clean, useful data from the pipelines. **Please
[read the release notes](https://github.com/CAVaccineInventory/vaccine-feed-ingest-schema/releases/tag/1.0.0)
to understand what changed.**

This change was pre-announced in #278, and is likely to break most (if
not all) currently written normalizers.

Additionally, remove the local copy of the schema, which was previously
deprecated in #225. 

BREAKING CHANGE: Validation enforced in schema.